Move CSVWriter tests into a separate module
